refactor(metrics): split metrics.rs god object into metrics/{onchain,http}
First paydown against the module-size ratchet: metrics.rs (786 LOC) → metrics/
onchain.rs (on-chain counters) + http.rs (per-endpoint HTTP metrics) + mod.rs
(wiring + tests), each under the 600 threshold. Pure code motion — 19 metrics
tests green. Baseline updated to drop the entry.
