+++ scripts/.god-object-baseline.txt
@@ -1,6 +1,6 @@
 # God-object baseline (LOC ceilings). Files here may only SHRINK.
-# Regenerate ONLY when a file legitimately drops: scripts/check-file-sizes.sh --update
-# The doctrine (CLAUDE.md 'Module Size Discipline'): split files by responsibility toward <600 LOC.
+# Regenerate after a real split: scripts/check-file-sizes.sh --update
+# Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
 10003 sandbox-runtime/src/operator_api.rs
 5385 sandbox-runtime/src/runtime.rs
 2966 ui/src/routes/create.tsx
@@ -16,7 +16,6 @@
 922 ui/src/routes/workflows._index.tsx
 884 sandbox-runtime/src/tee/verify.rs
 831 sandbox-runtime/src/reaper.rs
-786 sandbox-runtime/src/metrics.rs
 744 sandbox-runtime/src/scoped_session_auth.rs
 721 ui/src/components/shared/SessionSidebar.tsx
 701 ui/src/lib/stores/chatSessions.ts
